#79 fix: competitionController 파일을 일반, 디테일, 결과 파일로 분배
